Plenty of stuff
Some html and css. Some JSON parsing. Some helper modules.

cgi_helper.py is not intended to be used a great deal. It generates
html. I will mainly be using templating.

login.html, landing.html, and mekong.html are my templates, they are
not (for now) very templaty, but that will come.

mekong.css is the css for the website, I'll make it good later.

pages.py is a helper module for the templating. It has functions that
load specific html pages, and will, in future, also contain most of the
templating.