COOOKIIIIES
Users now have sessions, stored in the users database. A session has an
id, which is a random uuid, and are managed with cookies.

If a user is logged in, things will change. The username will replace
"login or register", an "add to cart" button (currently non-functional)
will become clickable on book detail page.